:recycle: Refactor wasm layout functions to their own submodule
