+++ render-wasm/src/wasm.rs
@@ -1,5 +1,6 @@
 pub mod fills;
 pub mod fonts;
+pub mod layout;
 pub mod paths;
 pub mod strokes;
 pub mod text;
